requirements.txt
paho-mqtt==1.5.1
requests==2.22.0
prometheus_client==0.7.1